+++ src/components/App/LaunchScreen/LaunchScreen.tsx
@@ -1,11 +1,14 @@
 import styled from 'styled-components'
 import {LogoImage} from "../../UI/Logo/Logo";
+import Loader from "react-loader-spinner";
+import React from "react";
 
 const StyledContainer = styled.div`
   height: 100%;
   width: 100%;
   display: flex;
-  background-color: ${props => props.theme.ui.colors.primary};
+  background-color: ${props => props.theme.ui.colors.secondary};
+  padding: 50px;
   flex-direction: column;
   align-items: center;
   justify-content: center;
@@ -15,6 +18,12 @@ const LaunchScreen = () => {
     return (
         <StyledContainer>
             <LogoImage logoColor='white'/>
+            <Loader
+                type="Oval"
+                color="white"
+                height={40}
+                width={40}
+            />
         </StyledContainer>
     )
 }