+++ src/components/App/PlatformModal/ManageProfiles/ManageProfiles.tsx
@@ -6,7 +6,7 @@ import ListItem from "../../../UI/Common/ListItem/ListItem";
 import Image from "../../../UI/Common/Profile/Image/Image";
 import AnonymousAvatar from '../../../../assets/Avatar-anonymous.png'
 import {TippyDropdown} from "../../../UI/Tippy/Tippy";
-import {ReactNode, useState} from "react";
+import {useState} from "react";
 import {ChevronLeft, VerticalOptions} from "../../../UI/Icons/Icons";
 import {PageNavigatorProvider} from "../../../Providers/PageNavigatorProvider";
 import {usePageNavigator} from "../../../../contexts/pageNavigator/PageNavigatorContext";
@@ -17,8 +17,8 @@ import ManageProfile from "./Pages/ManageProfile";
 import DeleteProfile from "./Pages/DeleteProfile";
 import {pageNavigator} from "../../../../contexts/pageNavigator/pageNavigatorActions";
 import {FormattedMessage} from "react-intl";
-import {ProfileData} from "../../../../store/appState/appStateReducer";
 import ContextMenu, {ContextMenuActions} from "../../../UI/Common/ContextMenu/ContextMenu";
+import {ProfileData} from "../../../../store/appState/appStateReducer";
 
 const StyledManagesProfileContainer = styled.div`
   width: 100%;